Daily observers update
When I tried to print out the list of zones worked I found the were
some duplicate zones. So I used a rather scruffy way to clean this
up (convert to and from int) and then I added code to print out the
zones. I also added code to sort the list of Maidenhead squares.